% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationmethods.R
\name{plot.effectsMed}
\alias{plot.effectsMed}
\title{Plot function for objects of class \code{"effectsMed"}}
\usage{
\method{plot}{effectsMed}(
  x,
  effect = "indirect",
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  lwd = graphics::par("lwd"),
  ...
)
}
\arguments{
\item{x}{object of class \code{"effectsMed"}}

\item{effect}{which effect to plot results for ("indirect" or "direct")}

\item{xlab}{a title for the x axis, see \code{\link{title}}. Default is \code{expression(rho)}.}

\item{ylab}{a title for the y axis, see \code{\link{title}}. Default is \code{NIE} (\code{NIE*} if \code{alt.decomposition = TRUE}) or \code{NDE} (\code{NDE*})}

\item{xlim}{the x limits (x1, x2) of the plot, see \code{\link{plot.default}}. Default is \code{c(min(x$Rho),max(x$Rho))}}

\item{ylim}{the y limits of the plot. Default is \code{c(min(x$CI$CI.nie[,1]),max(x$CI$CI.nie[,2]))}}

\item{main}{a main title for the plot, see \code{\link{title}}}

\item{lwd}{line widths for the lines of the plot, see \code{\link{par}}}

\item{...}{additional graphical parameters to be passed to plotting functions, see \code{\link{par}}}
}
\description{
Plots the estimated natural indirect or direct effects with confidence intervals over the range of the sensitivity parameter \code{Rho}.
}
