% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediationmethods.R
\name{print.summaryeffectsMed}
\alias{print.summaryeffectsMed}
\title{Print function for objects of class \code{"summaryeffectsMed"}}
\usage{
\method{print}{summaryeffectsMed}(x, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
\item{x}{object of class \code{"summaryeffectsMed"}}

\item{digits}{number of digits to be printed.}

\item{...}{additional arguments}
}
\description{
Print function for objects of class \code{"summaryeffectsMed"}
}
